coin read api
