+++ crates/sui-json-rpc/src/lib.rs
@@ -22,6 +22,7 @@ use crate::metrics::MetricsLayer;
 
 pub mod api;
 pub mod bcs_api;
+pub mod coin_api;
 pub mod event_api;
 mod metrics;
 pub mod read_api;
